requirements.txt
python-dotenv==1.2.1
requests==2.32.5
pandas==2.3.3
apscheduler==3.11.1
gspread==6.2.1
gspread_dataframe==4.0.0
oauth2client==4.1.3